Improve the bootstrap scripts
This commit modifies the bootstrap scripts to download k8s
componentes only if the do not exist on the system.

In addition, when we build a pipeline for building the images
we don't want to configure k8s componentes, we just want to pre-fetch
the binaries.
Hence we can now call each of the bootstrap scripts with -f or --fetch
which will then only download the binaries and exist.