Working addition of Sec. group and rules
It's seems the the cause of the problem was close port 80.
For the sake of simplicity I removed async code here, to ease debugging.
Also, I replaced the openstack client with neutron, because I was not
sure the the former does what I need as I need. I _think_ this could
work both with async and openstack client, but I still have not tested
it.
It would be the simplest to open port 80 on the sec_group branch and see
if this works there too.
There is definitely a lot of potential of improving the security group
and the rules.
But for now it does the job, a small step to prevent operator error
forgeting to apply those rules.