Pipeline: implement a layman's lock
gitlab-ci is still a baby compared to GoCD and Jenkins.
Hence, it does not support a FIFO queue for jobs. Running multiple jobs
will exshaust our resources, and openstack will break our code.
To work around it we could:
 - increase openstack limit indefinitely.
 - check that no other job is running.

The first solution, is terrible and costs a lot of money.
The second solution is better. The number of volumes in openstack is a
good proxy to the number of server and hence to whether a pipeline is
currently active or not. Our test cluster has 6 machines, so if there is
one pipeline job active the number of volumes is greater than 5 and the
job will wait.