Pipeline: a FIFO
Until now, the wait script only looked of another job is running. This
is good enough for one developer working late night. As soon as more
people start pushing to gitlab, we get a gridlock. So this is an attept
to make the first scheduled to job to run first. Jobs have an ID an the
smallest one should run first

[ci skip]