Wait for each etcd to start
Before bootstrapping the next master we have to wait for etcd to start,
join and the vote should complete.
Earlier we started etcd-2, and immidietly rushed to bootstrap etcd-3,
which caused both to crash and kill the cluster.
Waiting for each to join is a bit slower but MUCH safer.