Make ssh more robust
Until now we made the assumption the hosts where known or just typed
yes when ssh warned us. If this script should run fully automated, it
would hang on this prompt.
Therefore, we added this options to avoid the prompt and also handle
connection refused from the server if it is not completely up yet.

[ci skip]