Improve bootstrap script for masters
Sometimes the cluster build failed, because we didn't properly
distribute all the keys and certificates.

This commit changes a bit the order of thing.

First, downloading calico can take a while, but it does not need
to block. Second, we distribute the keys right before we bootstrap
a master. Which gaurantess a master will successfully join the
cluster.