Increment the number of nodes in the cluster config
This commit changes the number of nodes found in the config
file. Assuming that the configuration represent the true state
the number of `n-nodes` is incremented in the config file
by the `--amount` given in the command line.

The configuration is then written to a yaml file and the user
is being notified about the new file name.