Refactor the dns-check, fix a false positiv
Previously, if the DNS failed, we didn't know about it. It warned but it
didn't fail. This is not desired since a cluster without DNS is
crippled.

In order to fix the false positive, I need to move the commands out from
the Makefile to shell script, since I encountered the limits of what can
be done conviniently in a Makefile