The dns-check can run in parallel to LB check
A cluster can have it's traffic from Node to Pod working, which
will result in a behaviour which external LB will work.
However, if a Pod to Pod communitication fails, DNS will fail.

Hence, these two tests are not dependant on each other and can run
in parallel, which make the pipeline run a bit faster.