Seperate example-config and pi-dev-config files
A new folder configs/ has been created on project root with two files in
it: example-config.yml and pi-dev-config.yml . The first is a general
reference file which should work out-of-the-box in conjunction to the
explanation of usage.rst . The second one is a sample configuration file
which should work out-of-the-box when developing for koris inside the PI
OpenStack project.

Furthermore, links inside README.rst and usage.rst have been adjusted so
they point to the new config files.