Increased delay on resource creation, fixed healthmonitor delete check
Due to an upgrade to OpenStack Pike, the time to create LoadBalancers seems
to have increased, as well as the "waiting" state is now being indicated
by an NeutronConflict exception. For this, NeutronConflict has been
added to the retry decorator for the functions that create listeners,
pools, healthmonitors and members. Also the delay time has been
increased so resources get created properly.

A fix has been introduced to properly run a `koris
destroy` in case no healthmonitor is present.

Due to the changing NeutronConflict, no pool members could been added
(the exception was `pass`ed previously). This has been changed and the
NeutronConflict is now being retried while trying to add members to a
pool.