more fixes for the bootstrap script
This makes the script more robust, removes hard coded ubuntu user and
allows bootsrapping all nodes if a list of worker nodes is given.

All dependencies are install on each master and node during run time,
since we can't assume that the koris image is available in VMware or
baremetal.

KORIS-117