Fix parsing of etcd cluster status
Now the pod will start with the correct args, but ssh to the new
master still fails.

Checking manually seems to work:

```
/ # hostname
master-adder
/ # ssh -i /etc/ssh/ssh_host_rsa_key -l ubuntu master-6-am
Welcome to Ubuntu 16.04.6 LTS (GNU/Linux 4.4.0-1041-kvm x86_64)

 * Documentation:  https://help.ubuntu.com
 * Management:     https://landscape.canonical.com
 * Support:        https://ubuntu.com/advantage

  Get cloud support with Ubuntu Advantage Cloud Guest:
    http://www.ubuntu.com/business/services/cloud
This system has been minimized by removing packages and content that are
not required on a system that users do not log into.

To restore this content, you can run the 'unminimize' command.

18 packages can be updated.
5 updates are security updates.

New release '18.04.2 LTS' available.
Run 'do-release-upgrade' to upgrade to it.

Last login: Tue Apr  2 21:40:13 2019 from 10.32.192.43
```

So, this is probably just a tiny issue.