Make abort pipeline smarter and cancel running for dev
When we finish the release and push the tag for the branch to the
dev branch gitlab will start a third job for the same commit.
Thus, we now have again a 2nd job for the release (previously cancelled
on the master).
This commit detects this and cancel the running job for both.
Running the target shows that the matching jobs are:

Caneclled pipeline 12097 -> branch dev
Caneclled pipeline 12095 -> branch dev
Caneclled pipeline 12093 -> branch master

The actual important job, for the tag isn't affected.