Implement Singleton class for Logging
Introduces a Singleton class which acts as a metaclass for logging.

The class `koris.util.logger.Logger` has this Singleton class as metaclass which ensures
that only a single object is created of it. The class also has a check that if multiple loggers
of the same name get instantiated, only a single handler gets added and not duplicate ones.

Functionality can be tested with unit tests and the output verified with:

```
py.test -s -x tests/test_logger.py
```

The logging level can be set via the command line, taking both strings and integers:

```
--verbosity {0,1,2,3,4,quiet,error,warning,info,debug}, -v {0,1,2,3,4,quiet,error,warning,info,debug}
set the verbosity level (0 = quiet, 1 = error, 2 =warning, 3 = info, 4 = debug) (default: 3)
```

The log levels are oriented on `kubectl`'s log levels which unfortunately are inverse to those in
Python.

Calling, for example, `Logger.info()` is a wrapper which adds some colors/formatting (can be disabled
with
the `color=False` argument to each function) and then calls Pythons standard `info()` logging function.
Having this wrapper allows us greater flexibility in in what string formats we support for logging.

Addtionally, the logging functions `success` and `question` have been added:

- `success` calls `info` in green with `[+]`-prefix
- `question` calls `print` in light blue with `[?]`-prefix. This is because we want user prompts always
  be visible.

Like this, further output with custom formatting can easily be added to the common Logger object.
Additionally, extensive source code documentation is added.