Capture output of bootstrap master as debug, fix CLI test
The output of the shell call in `koris.deploy.k8s.K8S.run_add_script` has
been moved into a `LOGGER.debug()` statement. The upside is that
during execution on `INFO` level, the output has been cut down
tremendously. The downside is that capturing the output is buffered
so the output is printed as a whole instead of line by line on
execution. (I have tried capturing it line by line and while it's
possible, complexity increases greatly, the ouput is not very nice
and the method to do it is deprecated)

Also included additional info statements, as well as syntactical
changes.

The unit test for the cli (`tests/test_cli.py`) has been adapted
(`shell=True` is the wrong way to go when passing an array as
command to `Popen`)