Add metrics server plugin
This commit adds a manifest with all k8s resources needed to start
the metrics server. Currently, the metrics server is always on and
there is no user exposed way to disable the plugin.
If desired the user can use kubectl to delete the metrics server post
installation.
Also, we bump the version of the kubernetes python client.