Fix for missing external networks
Due to recent changes in out openstack the builds for stable
koris version failed with an error:

No external network found

This fixes this issue. The solution chosen here is first to change
ext01 to bgp-noris. Also, if this is not found, it could be detected
from the config and finally, if there is an external network, it will be
chosen.
Only in the rarest case, the code should stop with `No external network
found`.