Batch update the loadbalancer
Instead of updating each master and waiting we use a public
API of Openstack to batch update the listener
The API is exposed always but the Python client does not have
a public function for it. Hence, the use of do_request directly.

Using this approach we can boot the cluster more quickly. In earlier
versions, with every master becoming ready, there was a long wait
for openstack to reconfigure the loadbalancer before a new master
can be added.
This also applies for the Nginx ingress. There was again a long wait
adding each LB member. Bulk adding all the nodes is much quicker.

If the batch update does not succeed for some reason, we fallback
to the old mechanism of adding each master separately to the
loadbalancer.