Fix kubernetes target for added master
Make the python code explicit about variables passed to the method.
The more important change is in the bootstrap script, which worked util
now because we always matched the test version to the image version.
But if the image has a newer kubernetes version this test would fail,
because the kubernetes version was not downgraded.